\name{show.prp.palettes}
\alias{show.prp.palettes}
\title{Show the built-in \code{prp} palettes.}
\description{
Display a diagram showing the built-in palettes
accepted by \code{prp}'s \code{box.palette} argument.
%
% (The \code{box.palette} argument also accepts \code{"auto"}
% and palettes specified as a vector of colors.)
}
\usage{
show.prp.palettes()
}
\examples{
show.prp.palettes()

}
